move oninstalled and content scripts
